requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
gunicorn>=22,<26
requests>=2.31,<3
jinja2>=3.1,<4
pandas>=2.1,<3
akshare>=1.14
sqlalchemy>=2.0,<3
python-jose[cryptography]>=3.3,<4
passlib[bcrypt]>=1.7,<2
bcrypt>=4.0,<5
cachetools>=5.3,<6
aiohttp>=3.9,<4
apscheduler>=3.10,<4
python-dotenv>=1.0  # 可选：自动加载 .env 文件
python-multipart>=0.0.6  # FastAPI 文件上传依赖
rapidocr-onnxruntime==1.2.3  # OCR 识别引擎（模型内置，无需网络下载）
Pillow>=10.0  # 图片预处理